requirements.txt
poco==0.96.9
pocoui==1.0.79
numpy==1.14.5
xlrd==1.2.0
airtest==1.1.3
aircv==1.4.6
urllib3==1.25.11
openpyxl==3.0.5
allure-pytest==2.8.18
allure-python-commons==2.8.18